refactor: 文件管道移至单独文件
